Add resource for get '/jobs' and add control flow for query_string parameters
