+++ app/models/job.rb
@@ -3,4 +3,5 @@ class Job < ApplicationRecord
   belongs_to :company
   belongs_to :category
   has_many :comments
+  scope :sort_by_location, -> {order(:city)}
 end
