+++ config/routes.rb
@@ -9,6 +9,8 @@ Rails.application.routes.draw do
     resources :comments
   end
 
+  get "/jobs", to: 'jobs#index'
+
   resources :categories
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
