+++ spec/features/jobs/user_sees_all_company_jobs_spec.rb
@@ -1,0 +1,16 @@
+require 'rails_helper'
+
+describe "User sees all jobs" do
+  scenario "a user sees all the jobs for a specific company" do
+    company = Company.create!(name: "ESPN")
+    category = Category.create(title: "Much Sportnstuf")
+    company.jobs.create!(title: "Developer", level_of_interest: 70, city: "Denver", category_id: category.id)
+    company.jobs.create!(title: "QA Analyst", level_of_interest: 70, city: "New York City", category_id: category.id)
+
+    visit company_path(company)
+
+    expect(page).to have_content("ESPN")
+    expect(page).to have_content("Developer")
+    expect(page).to have_content("QA Analyst")
+  end
+end
