+++ spec/features/jobs/user_sees_all_jobs_spec.rb
@@ -1,16 +1,16 @@
-require 'rails_helper'
+require "rails_helper"
 
-describe "User sees all jobs" do
-  scenario "a user sees all the jobs for a specific company" do
+describe "User sees all jobs in database" do
+  it "allows user to view jobs sorted by location" do
     company = Company.create!(name: "ESPN")
     category = Category.create(title: "Much Sportnstuf")
-    company.jobs.create!(title: "Developer", level_of_interest: 70, city: "Denver", category_id: category.id)
+    company.jobs.create!(title: "Developer", level_of_interest: 60, city: "Denver", category_id: category.id)
     company.jobs.create!(title: "QA Analyst", level_of_interest: 70, city: "New York City", category_id: category.id)
+    company.jobs.create!(title: "Intern", level_of_interest: 80, city: "Boston", category_id: category.id)
 
-    visit company_path(company)
+    visit "/jobs?sort=location"
 
-    expect(page).to have_content("ESPN")
-    expect(page).to have_content("Developer")
-    expect(page).to have_content("QA Analyst")
+    expect(page).to have_tag("ul:first-child", text: "Boston")
+    expect(page).to have_tag("ul:last-child", text: "New York City")
   end
 end
